First commit of qf-glance.vim
A way to glance at quickfix results. Press space on a quickfix result to
open it in a preview window. Press <return> to open the result for
editing. Press q to close the quickfix window and all other result
buffers. Potentially this will be moved out in an independent plugin.